requirements.txt
bcrypt==3.2.0
click==8.0.3
dnspython==2.1.0
email-validator==1.1.3
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
python-dotenv==0.19.2
requests==2.26.0
SQLAlchemy==1.4.27
virtualenv==20.10.0
Werkzeug==2.0.2
WTForms==3.0.1
MarkupSafe==1.1.1